(#6458) - consistent error handling for mapreduce
CouchDB 2 improves mapreduce error reporting in the face of invalid
values being passed to a built-in reduce function.

Previously, a query to a view that couldn't be reduced would result
in a 500 error. In 2.x, CouchDB returns a 200 response and contains
details of the error in the row.

This patch detects the presence of this error message and converts it
into a client-side error, emulating the CouchDB 1.x behaviour.